fasterrcnn/requirements.txt
_libgcc_mutex             0.1
absl-py                   0.14.1
adversarial-robustness-toolbox 1.8.0
albumentations            1.0.3
argon2-cffi               20.1.0
astor                     0.8.1
astunparse                1.6.3
async-generator           1.10
attrs                     20.3.0
backcall                  0.2.0
bleach                    3.3.0
ca-certificates           2021.1.19
cached-property           1.5.2
cachetools                4.2.4
certifi                   2020.12.5
cffi                      1.14.5
charset-normalizer        2.0.5
clang                     5.0
colorama                  0.4.4
contextlib2               21.6.0
cryptography              35.0.0
cupy-cuda112              9.4.0
cycler                    0.10.0
cython                    0.29.24
dataclasses               0.8
decorator                 4.4.2
defusedxml                0.7.1
docutils                  0.17.1
easydict                  1.9                      
entrypoints               0.3
fastrlock                 0.6
flatbuffers               1.12
gast                      0.4.0
google-auth               1.35.0
google-auth-oauthlib      0.4.6
google-pasta              0.2.0
grpcio                    1.41.0
h5py                      3.1.0
idna                      3.2
imageio                   2.9.0
importlib-metadata        3.7.3
ipykernel                 5.5.0
ipython                   7.16.1
ipython-genutils          0.2.0
ipywidgets                7.6.3
jedi                      0.18.0
jeepney                   0.7.1
jinja2                    2.11.3
joblib                    1.1.0
jsonschema                3.2.0
jupyter                   1.0.0
jupyter-client            6.1.12
jupyter-console           6.4.0
jupyter-core              4.7.1
jupyterlab-pygments       0.1.2
jupyterlab-widgets        1.0.0
keras                     2.6.0
keras-applications        1.0.8
keras-preprocessing       1.1.2
keyring                   23.2.1
kiwisolver                1.3.1
ld_impl_linux-64          2.33.1
libffi                    3.3
libgcc-ng                 9.1.0
libstdcxx-ng              9.1.0
llvmlite                  0.36.0
lxml                      4.6.3
markdown                  3.3.4
markupsafe                1.1.1
matplotlib                3.3.4
mistune                   0.8.4
nbclient                  0.5.3
nbconvert                 6.0.7
nbformat                  5.1.2
ncurses                   6.2
nest-asyncio              1.5.1
networkx                  2.5.1
notebook                  6.3.0
numba                     0.53.1
numpy                     1.19.5
nvidia-cublas             11.5.1.101
nvidia-cuda-nvrtc         11.3.58
nvidia-cuda-runtime       11.3.58
nvidia-cudnn              8.2.0.41
nvidia-pyindex            1.0.8
nvidia-tensorrt           8.0.0.3
oauthlib                  3.1.1
onnx                      1.4.1
opencv-python             4.5.1.48
opencv-python-headless    4.5.3.56
openssl                   1.1.1j
opt-einsum                3.3.0
packaging                 20.9
pandas                    1.1.5
pandocfilters             1.4.3
parso                     0.8.1
pexpect                   4.8.0
pickleshare               0.7.5
pillow                    8.1.2
pip                       21.0.1
pkginfo                   1.7.1
prometheus-client         0.9.0
prompt-toolkit            3.0.18
protobuf                  3.15.8
ptyprocess                0.7.0
pyasn1                    0.4.8
pyasn1-modules            0.2.8
pycocotools               2.0.2
pycparser                 2.20
pygments                  2.8.1
pyparsing                 2.4.7
pyrsistent                0.17.3
python                    3.6.13
python-dateutil           2.8.1
pytz                      2021.1
pywavelets                1.1.1
pyyaml                    5.4.1
pyzmq                     22.0.3
qtconsole                 5.0.3
qtpy                      1.9.0
readline                  8.1
readme-renderer           30.0
requests                  2.26.0
requests-oauthlib         1.3.0
requests-toolbelt         0.9.1
rfc3986                   1.5.0
rsa                       4.7.2
scikit-image              0.17.2
scikit-learn              0.24.2
scipy                     1.5.4
secretstorage             3.3.1
send2trash                1.5.0
setuptools                58.2.0
six                       1.15.0
sqlite                    3.35.1
style                     1.1.0
tensorboard               2.7.0
tensorboard-data-server   0.6.1
tensorboard-plugin-wit    1.8.0
tensorboardx              2.4
tensorflow-estimator      2.6.0
tensorflow-gpu            2.6.0
tensorflow-object-detection-api 0.1.1
termcolor                 1.1.0
terminado                 0.9.3
testpath                  0.4.4
threadpoolctl             3.0.0
tifffile                  2020.9.3
tk                        8.6.10
torch                     1.9.1
torchaudio                0.9.1
torchvision               0.10.1
tornado                   6.1
tqdm                      4.62.2
traitlets                 4.3.3
twine                     3.4.2
typing                    3.7.4.3
typing-extensions         3.7.4.3
update                    0.0.1
urllib3                   1.26.6
wcwidth                   0.2.5
webencodings              0.5.1
werkzeug                  2.0.2
wheel                     0.36.2
widgetsnbextension        3.5.1
wrapt                     1.12.1
xz                        5.2.5
zipp                      3.4.1
zlib                      1.2.11
